Adding feature columns.
The Bootstrap builds on a "grid system". Just like a layout grid in print design, it defines how many columns are available to place elements. It also defines inner and outer spacings.

The Bootstrap grid requires at least two DOM elements, which get a certain class, to work: a row and a column element.
- An element with the class "row" is a wrapper for a column. A row has always space for 12 columns.
- An element with the class "col-*" wraps a grid element inside the current row. A column can use from 1 to 12 columns of the row.

For the "col-*" element you can also define it's behaviour on different screen sizes. The last part of the class (e.g. "-4") defines the number of columns the element will use inside the 12 column row. So the ".col-lg-4" element will take up 1/3 (4/12) of the row. The second part of the class (e.g. "-lg-") defines how wide the viewport of the browser needs to be, to apply the class and column value. In our case the columns are only applied to screens that are wider than "lg" (a "breakpoint" defined by Bootstrap, e.g. a screen ≥ 992px). So, if the screen is narrower than 992px the column element will take up all 12 columns (because no width class is applied). If it is equal or larger than 992px it will take up the given columns (4 in our case). If you add more elements to the row, than would have space (e.g. adding a fourth "col-lg-4" element to our example), they will automatically wrap on a new line.

Test in your browser how the columns align on different screen sizes.
Also try to play around with the breakpoint value (xs, md, lg, xl) and column width (1 - 12).

You can find more information about the available breakpoints on https://getbootstrap.com/docs/4.3/layout/grid/#grid-options