Limit the width of the content.
Until now, our first "row" uses the whole width of the viewport. On a large screen this is not very convenient, because the columns become too wide. If you take a closer look, the content even slightly overflows the window horizontally. This is because of the way the grid layout uses inner and outer spacings to create a gutter between the columns.

For this purpose there is a third grid related element class: "container". A div with the class "container" will limit the max-width of the inner content to

To fix the overflow issue and limit the with of the content there is a third grid related element class: "container" (there is also a "container-fluid" class, which would only remove the overflow but still allow a width of 100%). The "container" class fixes the overflow issue and limits the max-width based on the current viewport. Additionally it will center the content when the viewport is wider than the max-width.

Test in your browser, how the container effects the row and columns on different screen sizes.