Introducing images.
Compared to text, images are not able to reflow. So be careful to test your layout, when you add images, because they may overflow your layout if there is not enough space.

You can test this by removing "width" and "height" from the "img" tag, which limits the size of the image. We will discover later on, how you can limit the width of the image to it's container.

The "rounded-circle" class is also provided by Bootstrap. It applies rounded corners to an element, so it appears as a circle. By default all elements in HTML are rectangles. "src" defines where the image should be loaded from. In this case we use a free mock image service. The "alt" tag is mandatory in HTML and defines a fallback description text in case the image can not be loaded.