A new row.
Time to add a new section to our layout. This time we use a different column for each width. So on this element the left column is slightly wider than the right one. Also note that we used a different breakpoint (md) for this columns.

The "hr" element before and after the "row" block adds a horizontal line (horizontal rule).

Test in the browser, how the new columns will need a smaller viewport before they stack on each other.