Filling in content.
Let's add some more text and images to our content. For the text we again use one of Bootstraps classes ("text-muted") to create a slight color difference in the title. The "lead" class on the "p" element creates a slightly bigger font. Also note, that we did not add a "width" or "height" value to the image but therefore used the "img-fluid" class. This class makes sure, that the image is as big as possible, but not bigger as the content (the "col-md-5" in our case). Note, that images are not scaled up, if they are to small.

Test in the browser how the image scales with the available space.