Extend content.
Now that we have the basic style for the new blocks, lets add some more of them. Did you note the additional "order-md-" class on the columns on the second element in the list? This class changes the order of the elements in the browser when the breakpoint is equal or larger than "md". This way we can make sure, that on mobile the text comes always before the image but on desktop the second element has an inverted column order. You can find more information about ordering classes on https://getbootstrap.com/docs/4.3/layout/grid/#order-classes

Test in the browser how "order-md-" effects the position of the elements.