Change the vertical alignment of text and image.
On desktop it would be nicer to align text and image on the vertical center. Bootstrap also offers a class for this case. With "align-items-lg-center" on the "row" we tell the browser to align text and image to the center when the viewport is wider than "lg". You can find out more about alignments of grid columns on https://getbootstrap.com/docs/4.3/layout/grid/#alignment

Test how the alignment changes on the different screen sizes.