Adding a footer.
For some additional links we place a footer at the bottom of our content. Instead of grid rows and columns we just use a "container" this time and use the "float-right" class to move the back to top to the right of the layout.

Test how the text wraps to give space to the "float-right" element on very small screens,