Bootstrap components.
Beside styling and layout classes Bootstrap also offers a set of components like navbars, progress bars, breadcrumbs, modals, etc. to add more functionality to a website. Let's add a navbar, which will contain the navigation for our website. You can find out more about the navbar component on https://getbootstrap.com/docs/4.3/components/navbar/

Note how the navbar stays at the top of the page, when you scroll but also covers parts of the main content when your on top of the page. Also test how the header behaves on different screen sizes. Currently it is not possible to use the navigation on mobile.