Extending the navbar.
As we already discovered the navigation is not available on mobile. This is because there is just not enough space. But Bootstrap has a solution for that: a burger menu.

The button "navbar-toggler" will be displayed on mobile devices and allows us to toggle the visibility of the navigation. The new "id" on "collapse" is an identifier we can use on the "navbar-toggler" to control the visibility of the navigation with it. What sounds like magic is actually some JavaScript provided by Bootstrap, that registers clicks on the toggle button and then changes the classes on the "collapse".

We also added a tiny form (an other component), which represents a site search. It is only "fluff" and has currently no functionality. The "mr-", "mr-", "my-" classes on the form elements set outer spacings (margins) on the element. You can find out more about Bootstraps spacing system on https://getbootstrap.com/docs/4.3/utilities/spacing/

Inspect the "collapse" in the browsers development tools and investigate how the classes on the "collapse" and "navbar-toggler" change when you click the burger menu.