Adding a hero banner.
The term hero banner is used for large images on top of a website. Since our layout could still use some more color we will add an image as hero banner to the head of the page. To make sure, that the image covers the whole with we need to define some additional styles. We also move part of the header spacing from the body to the "hero-banner".

Test how the new banner looks on different screen sizes.