Replacing the hero-banner with a carousel.
To be honest, our hero-banner is a bit boring. Why don't we replace it with a carousel, to bring something dynamic to our page. To do this, we can use the carousel component of Bootstrap. You can find more information about it on https://getbootstrap.com/docs/4.3/components/carousel/

"carousel-inner" contains a "div" for each slide we want to display. We did not use the images directly, because we may wan't to add other content to the carousel in the future, too. The two links below the carousel-items are the next and previous buttons of the carousel. Almost all styles are provided by Bootstrap.

Test how the carousel looks on different screen sizes.