Adding additional content to the carousel.
OK, the client eventually asked for text and links in the carousel. Luckily we already prepared for this case, when we wrapped our slide images in div's. Now we can just add additional elements with the required content. Note, that Bootstrap also gives us a handful of classes we can use to place the text content above the image. If you remove the "carousel-caption" class, the text would be displayed below the image. Unfortunately the text does not work well on smaller screens. We will fix that in the next step.

Can you find out how the alignment of the carousel text is controlled? Test in the browser, what happens when you disable the "carousel-caption" and "text-" classes on the content elements. If you want to stop the animation for a moment you can remove data-ride="carousel" on the "carousel" item.