Improving mobile layout of carousel.
On mobile the carousel height is quite low and the text starts to overflow. To fix this we will set a fixed height to the carousel items (and images). We also use a new CSS syntax at this point: a media query. Just like the Bootstrap classes we used on the grid, they help us to control when certain styles should be applied. Here we will increase the height of the carousel and the font-size of the caption when the screen is wider than 641px. You can find out more about media queries on https://developer.mozilla.org/de/docs/Web/CSS/Media_Queries/Using_media_queries

Test in the browser how the font-size and carousel height changes around 641px viewport width. Also note, that the text still overflows the content below 320px: one of the main problems with responsive layouts. How would you resolve this problem?